Improve scripts to only install things that haven't been installed before

+++ build_scripts/ubuntu_sim_nuttx.sh
@@ -3,21 +3,34 @@
 # Ubuntu Config
 sudo apt-get remove modemmanager -y
 
+
 # Ninja build system
-mkdir -p $HOME/ninja
-cd $HOME/ninja
-wget https://github.com/martine/ninja/releases/download/v1.6.0/ninja-linux.zip
-unzip ninja-linux.zip
-rm ninja-linux.zip
-exportline="export PATH=$HOME/ninja:\$PATH"
-if grep -Fxq "$exportline" ~/.profile; then echo nothing to do ; else echo $exportline >> ~/.profile; fi
-. ~/.profile
+ninja_dir=$HOME/ninja
+echo "Installing Ninja to: $ninja_dir."
+if [ -d "$ninja_dir" ]
+then
+    echo " Ninja already installed."
+else
+    pushd .
+    mkdir -p $ninja_dir
+    cd $ninja_dir
+    wget https://github.com/martine/ninja/releases/download/v1.6.0/ninja-linux.zip
+    unzip ninja-linux.zip
+    rm ninja-linux.zip
+    exportline="export PATH=$ninja_dir:\$PATH"
+    if grep -Fxq "$exportline" ~/.profile; then echo " Ninja already in path" ; else echo $exportline >> ~/.profile; fi
+    . ~/.profile
+    popd
+fi
+
 
 # Common Dependencies
+echo "Installing common dependencies"
 sudo add-apt-repository ppa:george-edison55/cmake-3.x -y
-sudo apt-get update -y
+sudo apt-get update
 sudo apt-get install python-argparse git-core wget zip python-empy qtcreator cmake build-essential genromfs -y
 # required python packages
+sudo apt-get install python-dev -y
 sudo apt-get install python-pip
 sudo -H pip install pandas jinja2
 
@@ -30,7 +43,7 @@ sudo sh -c 'echo "deb http://packages.osrfoundation.org/gazebo/ubuntu-stable `ls
 ## Setup keys
 wget http://packages.osrfoundation.org/gazebo.key -O - | sudo apt-key add -
 ## Update the debian database:
-sudo apt-get update -y
+sudo apt-get update
 ## Install Gazebo8
 sudo apt-get install gazebo8 -y
 ## For developers (who work on top of Gazebo) one extra package
@@ -46,27 +59,47 @@ sudo apt-get install python-serial openocd \
 sudo apt-get remove gcc-arm-none-eabi gdb-arm-none-eabi binutils-arm-none-eabi gcc-arm-embedded
 sudo add-apt-repository --remove ppa:team-gcc-arm-embedded/ppa
 
-# Install GCC 5.4
-pushd .
-cd ~
-wget https://developer.arm.com/-/media/Files/downloads/gnu-rm/5_4-2016q2/gccarmnoneeabi542016q220160622linuxtar.bz2
-tar -jxf gccarmnoneeabi542016q220160622linuxtar.bz2
-exportline="export PATH=$HOME/gcc-arm-none-eabi-5_4-2016q2/bin:\$PATH"
-if grep -Fxq "$exportline" ~/.bash_profile; then echo nothing to do ; else echo $exportline >> ~/.profile; fi
-popd
 
-# Install 32 bit support libraries (ignore if fails)
-sudo dpkg --add-architecture i386
-sudo apt-get update
-sudo apt-get install libc6:i386 libgcc1:i386 libstdc++5:i386 libstdc++6:i386
-sudo apt-get install gcc-4.6-base:i386
+
+# Install GCC 5.4
+gcc_dir=$HOME/gcc-arm-none-eabi-5_4-2016q2
+echo "Installing GCC to: $gcc_dir"
+if [ -d "$gcc_dir" ]
+then
+    echo " GCC already installed."
+else
+    pushd .
+    cd ~    
+    wget https://developer.arm.com/-/media/Files/downloads/gnu-rm/5_4-2016q2/gccarmnoneeabi542016q220160622linuxtar.bz2
+    tar -jxf gccarmnoneeabi542016q220160622linuxtar.bz2
+    exportline="export PATH=$HOME/gcc-arm-none-eabi-5_4-2016q2/bin:\$PATH"
+    if grep -Fxq "$exportline" ~/.profile; then echo " GCC path already set." ; else echo $exportline >> ~/.profile; fi
+    . ~/.profile
+    popd
+    
+    # Install 32 bit support libraries (ignore if fails)
+    sudo dpkg --add-architecture i386
+    sudo apt-get update
+    sudo apt-get install libc6:i386 libgcc1:i386 libstdc++5:i386 libstdc++6:i386
+    sudo apt-get install gcc-5.4-base:i386
+fi
+
 
 
 # Clone PX4/Firmware
-mkdir -p ~/src
-cd ~/src
-git clone https://github.com/PX4/Firmware.git
-cd Firmware
+clone_dir=~/src
+echo "Cloning PX4 to: $ninja_dir."
+if [ -d "$clone_dir" ]
+then
+    echo " Firmware already cloned."
+else
+    mkdir -p $clone_dir
+    cd ~/$clone_dir
+    git clone https://github.com/PX4/Firmware.git
+    cd Firmware
+fi
+cd $clone_dir/Firmware
+    
 
 #Reboot the computer (required before building)
 echo RESTART YOUR COMPUTER to complete installation of PX4 development toolchain
